📦 feat(package.json): add fontawesome packages for icons 🔥 remove(package.json): remove lucide package as it's replaced by fontawesome 🔧 refactor(myanimelist-logo.svg, trakt-logo.svg): set fixed width and height for SVGs 🔧 refactor(Content.astro): remove unnecessary console.log ✨ feat(font-awesome.script.astro): create script to import and use fontawesome icons
FontAwesome packages have been added to provide a wider range of icons and better customization options. The Lucide package has been removed as it's no longer needed. Fixed width and height have been set for the SVGs to ensure consistent sizing. An unnecessary console.log has been removed from Content.astro. A new script has been created to import and use FontAwesome icons, making it easier to manage and maintain the icons used in the project.

🔧 refactor(BaseLayout.astro): extract FontAwesome script import to a separate component
The FontAwesome script import has been moved to a separate component (FontAwesomeScript) to improve code organization and maintainability. This change makes it easier to manage and update the FontAwesome script import in the future.